Migrate to the `nordtheme` organization (#14)
As part of the "Northern Post - The state and roadmap of Nord" [2]
initiative, this repository will be migrated to the `nordtheme` GitHub
organization [1]. This included the temporary removal of documentation,
since this is outdated by far anyway, and the move to the `nordtheme`
organization as a kind of "test subject" in order to ensure that all
required metadata and information, e.g. stars and watchers, will be
migrated.
Note that the documentation will be recreated as part of the data
transition [3] later on again, following the new project structure and
layout conventions.

[1]: https://github.com/nordtheme
[2]: https://github.com/arcticicestudio/nord/issues/180
[3]: https://github.com/arcticicestudio/nord-docs/issues/133

GH-13